+++ test/helpers/process.ex
@@ -15,8 +15,15 @@ defmodule Commanded.Helpers.Process do
     assert_receive {:DOWN, ^ref, _, _, _}, 5_000
   end
 
-  def shutdown(aggregate_uuid) do
-    pid = apply(@registry_provider, :whereis_name, [{:aggregate_registry, aggregate_uuid}])
+  def shutdown(name) when is_atom(name) do
+    case Process.whereis(name) do
+      nil -> :ok
+      pid -> shutdown(pid)
+    end
+  end
+
+  def shutdown(aggregate_module, aggregate_uuid) do
+    pid = apply(@registry_provider, :whereis_name, [{:aggregate_registry, {aggregate_module, aggregate_uuid}}])
     shutdown(pid)
   end
 end